divide forwarding logic
